+++ NEWS.md
@@ -1,3 +1,15 @@
+# BirdFlowR 0.1.0.9032
+2023-09-20
+
+* BREAKING change to column names in data frame returned by `rasterize_distr(format = "dataframe")` 
+  * Old `"time"` is now `"label"` and is an ordered factor it is derived from
+  the column names in the distribution which are often but not always dates.
+  * Old `density` is now value. 
+  
+* New functions to visualize distributions:
+  * `plot_distr()` 
+  * `animate_distr()` 
+
 # BirdFlowR 0.1.0.9031
 2023-09-08